Fix DeepSource issues (#211)
* Implicit type for new expressions

* Use object/collection initialization

* Use auto properties

* Use expression for simple properties

* Use method group instead of lambda

+++ zzre/rendering/TextureAssetLoader.cs
@@ -87,7 +87,7 @@ namespace zzre
             {
                 fixed (void* dataBytePtr = image.Data)
                 {
-                    IntPtr dataPtr = new IntPtr(dataBytePtr);
+                    IntPtr dataPtr = new(dataBytePtr);
                     device.UpdateTexture(texture,
                         source: dataPtr,
                         sizeInBytes: (uint)image.DataLen,
